+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/DefaultAudioSink.java
@@ -684,7 +684,7 @@ public final class DefaultAudioSink implements AudioSink {
     if (!offloadDisabledUntilNextConfiguration && useOffloadedPlayback(format, audioAttributes)) {
       return SINK_FORMAT_SUPPORTED_DIRECTLY;
     }
-    if (isPassthroughPlaybackSupported(format, audioCapabilities)) {
+    if (audioCapabilities.isPassthroughPlaybackSupported(format)) {
       return SINK_FORMAT_SUPPORTED_DIRECTLY;
     }
     return SINK_FORMAT_UNSUPPORTED;
@@ -767,7 +767,7 @@ public final class DefaultAudioSink implements AudioSink {
         outputMode = OUTPUT_MODE_PASSTHROUGH;
         @Nullable
         Pair<Integer, Integer> encodingAndChannelConfig =
-            getEncodingAndChannelConfigForPassthrough(inputFormat, audioCapabilities);
+            audioCapabilities.getEncodingAndChannelConfigForPassthrough(inputFormat);
         if (encodingAndChannelConfig == null) {
           throw new ConfigurationException(
               "Unable to configure passthrough for: " + inputFormat, inputFormat);
@@ -1691,134 +1691,6 @@ public final class DefaultAudioSink implements AudioSink {
     return configuration.outputMode == OUTPUT_MODE_PCM
         ? (writtenPcmBytes / configuration.outputPcmFrameSize)
         : writtenEncodedFrames;
-  }
-
-  private static boolean isPassthroughPlaybackSupported(
-      Format format, AudioCapabilities audioCapabilities) {
-    return getEncodingAndChannelConfigForPassthrough(format, audioCapabilities) != null;
-  }
-
-  /**
-   * Returns the encoding and channel config to use when configuring an {@link AudioTrack} in
-   * passthrough mode for the specified {@link Format}. Returns {@code null} if passthrough of the
-   * format is unsupported.
-   *
-   * @param format The {@link Format}.
-   * @param audioCapabilities The device audio capabilities.
-   * @return The encoding and channel config to use, or {@code null} if passthrough of the format is
-   *     unsupported.
-   */
-  @Nullable
-  private static Pair<Integer, Integer> getEncodingAndChannelConfigForPassthrough(
-      Format format, AudioCapabilities audioCapabilities) {
-    @C.Encoding
-    int encoding = MimeTypes.getEncoding(checkNotNull(format.sampleMimeType), format.codecs);
-    // Check for encodings that are known to work for passthrough with the implementation in this
-    // class. This avoids trying to use passthrough with an encoding where the device/app reports
-    // it's capable but it is untested or known to be broken (for example AAC-LC).
-    boolean supportedEncoding =
-        encoding == C.ENCODING_AC3
-            || encoding == C.ENCODING_E_AC3
-            || encoding == C.ENCODING_E_AC3_JOC
-            || encoding == C.ENCODING_AC4
-            || encoding == C.ENCODING_DTS
-            || encoding == C.ENCODING_DTS_HD
-            || encoding == C.ENCODING_DOLBY_TRUEHD;
-    if (!supportedEncoding) {
-      return null;
-    }
-    if (encoding == C.ENCODING_E_AC3_JOC
-        && !audioCapabilities.supportsEncoding(C.ENCODING_E_AC3_JOC)) {
-      // E-AC3 receivers support E-AC3 JOC streams (but decode only the base layer).
-      encoding = C.ENCODING_E_AC3;
-    } else if (encoding == C.ENCODING_DTS_HD
-        && !audioCapabilities.supportsEncoding(C.ENCODING_DTS_HD)) {
-      // DTS receivers support DTS-HD streams (but decode only the core layer).
-      encoding = C.ENCODING_DTS;
-    }
-    if (!audioCapabilities.supportsEncoding(encoding)) {
-      return null;
-    }
-
-    int channelCount;
-    if (encoding == C.ENCODING_E_AC3_JOC) {
-      // E-AC3 JOC is object based so the format channel count is arbitrary. From API 29 we can get
-      // the channel count for this encoding, but before then there is no way to query it so we
-      // assume 6 channel audio is supported.
-      if (Util.SDK_INT >= 29) {
-        // Default to 48 kHz if the format doesn't have a sample rate (for example, for chunkless
-        // HLS preparation). See [Internal: b/222127949].
-        int sampleRate = format.sampleRate != Format.NO_VALUE ? format.sampleRate : 48000;
-        channelCount =
-            getMaxSupportedChannelCountForPassthroughV29(C.ENCODING_E_AC3_JOC, sampleRate);
-        if (channelCount == 0) {
-          Log.w(TAG, "E-AC3 JOC encoding supported but no channel count supported");
-          return null;
-        }
-      } else {
-        channelCount = 6;
-      }
-    } else {
-      channelCount = format.channelCount;
-      if (channelCount > audioCapabilities.getMaxChannelCount()) {
-        return null;
-      }
-    }
-    int channelConfig = getChannelConfigForPassthrough(channelCount);
-    if (channelConfig == AudioFormat.CHANNEL_INVALID) {
-      return null;
-    }
-
-    return Pair.create(encoding, channelConfig);
-  }
-
-  /**
-   * Returns the maximum number of channels supported for passthrough playback of audio in the given
-   * format, or 0 if the format is unsupported.
-   */
-  @RequiresApi(29)
-  private static int getMaxSupportedChannelCountForPassthroughV29(
-      @C.Encoding int encoding, int sampleRate) {
-    android.media.AudioAttributes audioAttributes =
-        new android.media.AudioAttributes.Builder()
-            .setUsage(android.media.AudioAttributes.USAGE_MEDIA)
-            .setContentType(android.media.AudioAttributes.CONTENT_TYPE_MOVIE)
-            .build();
-    // TODO(internal b/25994457): Query supported channel masks directly once it's supported.
-    for (int channelCount = 8; channelCount > 0; channelCount--) {
-      AudioFormat audioFormat =
-          new AudioFormat.Builder()
-              .setEncoding(encoding)
-              .setSampleRate(sampleRate)
-              .setChannelMask(Util.getAudioTrackChannelConfig(channelCount))
-              .build();
-      if (AudioTrack.isDirectPlaybackSupported(audioFormat, audioAttributes)) {
-        return channelCount;
-      }
-    }
-    return 0;
-  }
-
-  private static int getChannelConfigForPassthrough(int channelCount) {
-    if (Util.SDK_INT <= 28) {
-      // In passthrough mode the channel count used to configure the audio track doesn't affect how
-      // the stream is handled, except that some devices do overly-strict channel configuration
-      // checks. Therefore we override the channel count so that a known-working channel
-      // configuration is chosen in all cases. See [Internal: b/29116190].
-      if (channelCount == 7) {
-        channelCount = 8;
-      } else if (channelCount == 3 || channelCount == 4 || channelCount == 5) {
-        channelCount = 6;
-      }
-    }
-
-    // Workaround for Nexus Player not reporting support for mono passthrough. See
-    // [Internal: b/34268671].
-    if (Util.SDK_INT <= 26 && "fugu".equals(Util.DEVICE) && channelCount == 1) {
-      channelCount = 2;
-    }
-
-    return Util.getAudioTrackChannelConfig(channelCount);
   }
 
   private boolean useOffloadedPlayback(Format format, AudioAttributes audioAttributes) {